closes #23
