+++ app/blog/post/[slug]/page.tsx
@@ -1,10 +1,9 @@
 import React from "react";
-// import ReactMarkdown from "react-markdown";
 import BlogLayout from "../../../../components/BlogLayout/BlogLayout";
 import { getFiles, getPostBySlug } from "../../../../lib/utils";
-// import Link from "next/link";
 import { parseMarkdownToSections } from "../../../../lib/parseMarkdown";
 import Image from "next/image";
+import type { Section } from "../../../../lib/types";
 
 export default async function BlogPost({
   params,
@@ -13,42 +12,31 @@ export default async function BlogPost({
 }) {
   const { slug } = await params;
   const post = await getPostBySlug(slug);
-  const sections = parseMarkdownToSections(post.markdownBody);
 
   if (!post) {
     return <div>Post not found</div>;
   }
 
-  return (
-    // <BlogLayout post={post}>
-    //   <ReactMarkdown
-    //     components={{
-    //       a: ({ href, children }) => <Link href={href || "#"}>{children}</Link>,
-    //     }}
-    //   >
-    //     {post.markdownBody}
-    //   </ReactMarkdown>
-    // </BlogLayout>
+  const sections: Section[] = parseMarkdownToSections(post.markdownBody);
 
+  return (
     <BlogLayout post={post}>
-      <article className="prose">
+      <article className="prose prose-lg max-w-none">
         {sections.map((section, i) => {
           if (section.type === "text") {
-            return section.content.map((para, j) => (
-              <p key={`text-${i}-${j}`}>{para}</p>
-            ));
+            return <p key={`text-${i}`}>{section.content}</p>;
           }
 
           if (section.type === "image") {
             return section.content.map((url, j) => (
-              <div key={`img-${i}-${j}`} className="my-4">
+              <div key={`img-${i}-${j}`} className="my-6">
                 <Image
                   src={url}
                   alt=""
                   width={800}
                   height={500}
                   className="rounded-lg w-full h-auto object-cover"
-                />{" "}
+                />
               </div>
             ));
           }
